pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-emotion-recognition"
version = "1.0.0"
description = "Speech Emotion Recognition using TensorFlow, Keras, and Librosa"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Tharun Ponnam", email = "tharunponnam007@gmail.com"}
]
maintainers = [
    {name = "Tharun Ponnam", email = "tharunponnam007@gmail.com"}
]
keywords = [
    "speech emotion recognition",
    "deep learning",
    "tensorflow",
    "keras",
    "librosa",
    "audio processing",
    "machine learning",
    "affective computing"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
]
requires-python = ">=3.8"
dependencies = [
    "tensorflow>=2.10.0",
    "librosa>=0.10.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "soundfile>=0.12.0",
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
app = [
    "streamlit>=1.20.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "python-multipart>=0.0.6",
]
notebook = [
    "jupyter>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "ipywidgets>=8.0.0",
]
all = [
    "speech-emotion-recognition[dev,app,notebook]",
]

[project.urls]
Homepage = "https://github.com/tharun-ship-it/speech-emotion-recognition"
Documentation = "https://github.com/tharun-ship-it/speech-emotion-recognition#readme"
Repository = "https://github.com/tharun-ship-it/speech-emotion-recognition.git"
Issues = "https://github.com/tharun-ship-it/speech-emotion-recognition/issues"

[project.scripts]
ser-train = "scripts.train:main"
ser-evaluate = "scripts.evaluate:main"
ser-extract = "scripts.extract_features:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
